use longs for most things instead of integers

+++ src/wbs/applications/imchat/api/ImChatConversationEndRequest.java
@@ -16,6 +16,6 @@ class ImChatConversationEndRequest {
 	String sessionSecret;
 
 	@DataAttribute
-	Integer conversationIndex;
+	Long conversationIndex;
 
 }
